packages/Classification_backend/pyproject.toml
[project]
name = "classification-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "boto3>=1.40.14",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
